refactor(mobile): extract shared asset and preference types
