+++ apps/mobile/src/core/apis/address.ts
@@ -3,7 +3,7 @@ import { KEYRING_TYPE } from '@rabby-wallet/keyring-utils';
 import WatchKeyring from '@rabby-wallet/eth-keyring-watch';
 
 import { isSameAccount } from '@/hooks/accountsSwitcher';
-import { KeyringAccountWithAlias } from '@/hooks/account';
+import type { KeyringAccountWithAlias } from '@/types/account';
 import {
   contactService,
   dappService,
