+++ apps/mobile/src/core/apis/provider.ts
@@ -25,7 +25,7 @@ import abiCoder, { AbiCoder } from 'web3-eth-abi';
 import { IExtractFromPromise } from '@/utils/type';
 import { findChain } from '@/utils/chain';
 import { Tx } from '@rabby-wallet/rabby-api/dist/types';
-import { Account } from '../services/preference';
+import type { Account } from '@/types/account';
 import { isTempoChain } from '@/utils/tempo';
 import { decodeFunctionResult, encodeFunctionData } from 'viem';
 import { Abis as TempoAbis, Addresses as TempoAddresses } from 'viem/tempo';
