+++ apps/mobile/src/core/controllers/provider.ts
@@ -54,7 +54,7 @@ import RpcCache from '../services/rpcCache';
 import { CHAINS_ENUM } from '@/constant/chains';
 import { SAFE_RPC_METHODS } from '@/constant/rpc';
 import BaseController from './base';
-import { Account } from '../services/preference';
+import type { Account } from '@/types/account';
 import BigNumber from 'bignumber.js';
 // import { formatTxMetaForRpcResult } from 'background/utils/tx';
 import { findChain, findChainByEnum } from '@/utils/chain';
@@ -70,7 +70,7 @@ import { StatsData } from '../services/notification';
 import { ethers } from 'ethers';
 import { getGlobalProvider } from '../apis/globalProvider';
 import { bytesToHex } from '@ethereumjs/util';
-import { CustomTestnetTokenBase } from '../services/customTestnetService';
+import type { CustomTestnetTokenBase } from '@/types/customTestnet';
 // import { updateExpiredTime } from '@/databases/sync/assets';
 import { PENDGING_TIME } from '@/constant/expireTime';
 import { isString } from 'lodash';
