+++ apps/mobile/src/core/controllers/rpcFlow.ts
@@ -35,7 +35,7 @@ import { getAccountList } from '../apis/account';
 import { getDappAccount } from '@/hooks/useDapps';
 import { shouldAutoConnect, shouldAutoPersonalSign } from './autoConnect';
 import { openapi } from '../request';
-import { Account } from '../services/preference';
+import type { Account } from '@/types/account';
 
 export const resemblesETHAddress = (str: string): boolean => {
   return str.length === 42;
