+++ apps/mobile/src/core/services/customTestnetService.ts
@@ -33,43 +33,19 @@ import { updateChainStore } from '@/constant/chains';
 import { appStorage } from '../storage/mmkv';
 import { APP_STORE_NAMES } from '@/core/storage/storeConstant';
 import { matomoRequestEvent } from '@/utils/analytics';
+import type {
+  CustomTestnetToken,
+  CustomTestnetTokenBase,
+  TestnetChain,
+  TestnetChainBase,
+} from '@/types/customTestnet';
 
-export interface TestnetChainBase {
-  id: number;
-  name: string;
-  nativeTokenSymbol: string;
-  rpcUrl: string;
-  scanLink?: string;
-}
-
-export interface TestnetChain extends TestnetChainBase {
-  nativeTokenAddress: string;
-  hex: string;
-  network: string;
-  enum: CHAINS_ENUM;
-  serverId: string;
-  nativeTokenLogo: string;
-  eip: Record<string, any>;
-  nativeTokenDecimals: number;
-  scanLink: string;
-  isTestnet?: boolean;
-  logo: string;
-  whiteLogo?: string;
-  needEstimateGas?: boolean;
-  severity: number;
-}
-
-export interface CustomTestnetTokenBase {
-  id: string;
-  chainId: number;
-  symbol: string;
-  decimals: number;
-}
-
-export interface CustomTestnetToken extends CustomTestnetTokenBase {
-  amount: number;
-  rawAmount: string;
-}
+export type {
+  CustomTestnetToken,
+  CustomTestnetTokenBase,
+  TestnetChain,
+  TestnetChainBase,
+} from '@/types/customTestnet';
 
 export type CutsomTestnetServiceStore = {
   customTestnet: Record<string, TestnetChain>;
