+++ apps/mobile/src/hooks/account.ts
@@ -8,11 +8,15 @@ import React, {
 
 import { KeyringAccount, KEYRING_TYPE } from '@rabby-wallet/keyring-utils';
 
-import { Account, IPinAddress } from '@/core/services/preference';
+import type {
+  Account,
+  IPinAddress,
+  KeyringAccountWithAlias,
+} from '@/types/account';
 import { getWalletIcon } from '@/utils/walletInfo';
 import { filterMyAccounts } from '@/utils/account';
 import { useCreationWithShallowCompare } from './common/useMemozied';
-import { accountEvents, KeyringAccountWithAlias } from '@/core/apis/account';
+import { accountEvents } from '@/core/apis/account';
 import { resolveValFromUpdater, UpdaterOrPartials } from '@/core/utils/store';
 import addressBalanceStore from '@/store/balance';
 import accountStore, {
@@ -29,7 +33,7 @@ import { Store } from '@/core/services/hdKeyringService';
 import { perfEvents } from '@/core/utils/perf';
 import { AccountInfoEntity } from '@/databases/entities/accountInfo';
 
-export type { KeyringAccountWithAlias as /** @deprecated */ KeyringAccountWithAlias };
+export type { KeyringAccountWithAlias as /** @deprecated */ KeyringAccountWithAlias } from '@/types/account';
 
 export function useIsNewlyAddedAccount(account: KeyringAccount) {
   const dbId = useMemo(() => {
