+++ apps/mobile/src/store/account.ts
@@ -1,8 +1,4 @@
-import {
-  accountEvents,
-  fetchAllAccounts,
-  KeyringAccountWithAlias,
-} from '@/core/apis/account';
+import { accountEvents, fetchAllAccounts } from '@/core/apis/account';
 import { apiMnemonic } from '@/core/apis';
 import { getAllAccounts, removeAddress } from '@/core/apis/address';
 import { AccountInfoEntity } from '@/databases/entities/accountInfo';
@@ -12,7 +8,11 @@ import { deleteDBResourceForAddress } from '@/databases/sync/assets';
 import { BaseStore } from './_base';
 import { InteractionManager } from 'react-native';
 import { isEqual } from 'lodash';
-import { Account, IPinAddress } from '@/core/services/preference';
+import type {
+  Account,
+  IPinAddress,
+  KeyringAccountWithAlias,
+} from '@/types/account';
 import {
   keyringService,
   preferenceService,
