+++ apps/mobile/src/utils/sendTransaction.ts
@@ -48,7 +48,7 @@ import { getCexInfo } from '@/hooks/useCexSupportList';
 import { isNonPublicProductionEnv } from '@/constant';
 import { getDefaultStore } from 'jotai';
 import { mockBatchRevokeStore } from '@/hooks/appSettings';
-import { Account } from '@/core/services/preference';
+import type { Account } from '@/types/account';
 import miscService from '@/core/services/misc';
 import {
   buildTempoTransaction,
