+++ apps/mobile/src/utils/transaction.ts
@@ -21,7 +21,8 @@ import { findChain, getChain } from './chain';
 import i18n from './i18n';
 import { isTempoChain } from './tempo';
 import { openExternalUrl } from '@/core/utils/linking';
-import { Account, IManageToken } from '@/core/services/preference';
+import type { Account } from '@/types/account';
+import type { IManageToken } from '@/types/assets';
 import { HistoryItemEntity } from '@/databases/entities/historyItem';
 import { preferenceService, transactionHistoryService } from '@/core/services';
 import { CustomTxItem } from '@/core/services/transactionHistory';
